+++ src/cmd/tsv.rs
@@ -1,5 +1,5 @@
 use clap::*;
-use garr::*;
+use crate::*;
 use intspan::*;
 use redis::Commands;
 use std::collections::BTreeMap;