+++ CHANGELOG.md
@@ -8,7 +8,7 @@
   * `garr env` creates sqls
   * Stats of ctg and rsw
 
-* `garr stat` - await makes the compilation extremely slow
+* `garr stat` - `await` makes the compilation extremely slow
 
 ## 0.1.0 - 2021-08-27